Update README.md
Oasis Data Cleaning

Data cleaning involves identifying and rectifying errors, filling in missing information, and removing duplicates, which can hinder data analysis. As a skilled data analyst, my objective is to ensure that the data is cleaner for accurate analysis. This process provides a reliable foundation for decision-making and aligns with my goal of delivering precise and dependable insights. These insights are crucial for analyzing revenue, profit, loss, and other beneficial metrics to enhance business performance.